requirements.txt
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
Flask==3.0.3
Flask-Dance==7.1.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-secret-manager==2.22.0
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
PyJWT>=2.0.0,<3.0.0
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==24.2
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
urllib3==2.3.0
URLObject==2.4.3
Werkzeug==3.1.3
